# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved-places:
    get:
      operationId: listSavedPlaces
      tags: [saved-places]
      summary: List all saved places
      responses:
        "200":
          description: List of saved places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedPlace"
    post:
      operationId: createSavedPlace
      tags: [saved-places]
      summary: Save a new place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedPlaceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPlace"
  /saved-places/{id}:
    get:
      operationId: getSavedPlace
      tags: [saved-places]
      summary: Get a saved place by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPlace"
        "404":
          description: Not found
    put:
      operationId: updateSavedPlace
      tags: [saved-places]
      summary: Update a saved place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSavedPlaceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPlace"
    delete:
      operationId: deleteSavedPlace
      tags: [saved-places]
      summary: Delete a saved place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Trips ──────────────────────────────────────────────────────────────────
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip with its days and itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
    put:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Itinerary Items ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SavedPlace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        category:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        estimatedDuration:
          type: integer
          description: Estimated visit duration in minutes
          nullable: true
        estimatedCost:
          type: number
          description: Estimated cost in USD
          nullable: true
        priority:
          type: integer
          description: Priority 1-5
          nullable: true
        openingHours:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - address
        - latitude
        - longitude
        - createdAt
    CreateSavedPlaceBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        category:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        estimatedDuration:
          type: integer
          nullable: true
        estimatedCost:
          type: number
          nullable: true
        priority:
          type: integer
          nullable: true
        openingHours:
          type: string
          nullable: true
      required:
        - name
        - address
        - latitude
        - longitude
    UpdateSavedPlaceBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        category:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        estimatedDuration:
          type: integer
          nullable: true
        estimatedCost:
          type: number
          nullable: true
        priority:
          type: integer
          nullable: true
        openingHours:
          type: string
          nullable: true
    Trip:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
          nullable: true
        totalDays:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - startDate
        - endDate
        - totalDays
        - createdAt
    TripDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
          nullable: true
        totalDays:
          type: integer
        createdAt:
          type: string
          format: date-time
        days:
          type: array
          items:
            $ref: "#/components/schemas/TripDay"
      required:
        - id
        - name
        - startDate
        - endDate
        - totalDays
        - createdAt
        - days
    TripDay:
      type: object
      properties:
        dayNumber:
          type: integer
        date:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryItem"
      required:
        - dayNumber
        - date
        - items
    CreateTripBody:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
          nullable: true
      required:
        - name
        - startDate
        - endDate
    UpdateTripBody:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
          nullable: true
    ItineraryItem:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        savedPlaceId:
          type: integer
          nullable: true
        dayNumber:
          type: integer
        position:
          type: integer
        customName:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        estimatedDuration:
          type: integer
          nullable: true
        estimatedCost:
          type: number
          nullable: true
        place:
          $ref: "#/components/schemas/SavedPlace"
          nullable: true
      required:
        - id
        - tripId
        - dayNumber
        - position
...