# lib/api-zod/src/generated/types/createTripBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTripBody {
  name: string;
  startDate: Date;
  endDate: Date;
  description?: string | null;
}