# lib/api-zod/src/generated/types/updateSavedPlaceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSavedPlaceBody {
  name?: string;
  address?: string;
  latitude?: number;
  longitude?: number;
  category?: string | null;
  note?: string | null;
  estimatedDuration?: number | null;
  estimatedCost?: number | null;
  priority?: number | null;
  openingHours?: string | null;
}